Make all WorldLocations appear to be untranslated
This commit makes all WorldLocations appear to have no translations. We
are retiring the translated versions and need to prevent them being
published while the redirects are implemented and publishing API is
updated.

Once this is complete we can delete the data from
world_location_translations and revert this commit.