Add redirect checker
We'll be redirecting all worldwide related content from `/government/world` to `/world`. This checker class will test that the redirects are correct, by making a http request through the whole stack.
The idea is that you can add arrays of hashes with a `url` to check and the expected `redirect_url`.