Add WorldLocation redirect rake tasks
WorldLocations currently reside at the base_path
`/government/world/<slug>` and have translations that exist at
`/government/world/<slug>.<locale>`.

We are migrating WorldLocation pages to taxons which will exist at
`/world/<slug>` and will not support translations.

We are going to move the content in two steps.

Firstly we make all WorldLocations to appear to not have any
translations with the patch from the previous commit and then we
redirect the them to the :en version at `/government/world/<slug>`.
This will stop Publishing API presenting the translation links to content
store and they'll be cleaned up from content store. This commit
contains a rake task for that.

Once we have content at the new `/world/<slug>` paths there is
another rake task that will redirect all locales to the new `/world/<slug>`.

The actual translatable functionality and translated data will be
removed in subsequent PRs.